Fix horiz scroll pos and also wrapped cursor pos.
In single line mode, when typing "x" U+2714 U+FE0F "y" followed by 34
more "z":

1.  Typing a 39th letter scrolled and failed to fill the display line,
    and positioned the cursor incorrectly.
2.  Typing a 40th letter scrolled prematurely.  Typing a 41st letter
    got back into sync.

In multiline mode, when a grapheme partially fits and has to wrap to a
new line:

3.  Positioning the caret at the beginning of the grapheme caused the
    cursor to stay on the old line, instead of following the grapheme
    to the new line.

This change fixes all 3 issues, plus a benign "char32_t" typo, plus
fixes the wcwidth_iter width type to be int16_t.

Code mostly written by Codex, with direction and corrections.

+++ include/wcwidth.h
@@ -48,10 +48,10 @@ public:
     void            unnext();
     const char*     character_pointer() const { return m_chr_ptr; }
     uint32_t        character_length() const { return uint32_t(m_chr_end - m_chr_ptr); }
-    int32_t         character_wcwidth_signed() const { return m_chr_wcwidth; }
-    uint32_t        character_wcwidth_zeroctrl() const { return (m_chr_wcwidth < 0) ? 0 : m_chr_wcwidth; }
-    uint32_t        character_wcwidth_onectrl() const { return (m_chr_wcwidth < 0) ? 1 : m_chr_wcwidth; }
-    uint32_t        character_wcwidth_twoctrl() const { return (m_chr_wcwidth < 0) ? 2 : m_chr_wcwidth; }
+    int16_t         character_wcwidth_signed() const { return m_chr_wcwidth; }
+    uint16_t        character_wcwidth_zeroctrl() const { return (m_chr_wcwidth < 0) ? 0 : m_chr_wcwidth; }
+    uint16_t        character_wcwidth_onectrl() const { return (m_chr_wcwidth < 0) ? 1 : m_chr_wcwidth; }
+    uint16_t        character_wcwidth_twoctrl() const { return (m_chr_wcwidth < 0) ? 2 : m_chr_wcwidth; }
     bool            character_is_emoji() const { return m_emoji; }
     const char*     get_pointer() const;
     void            reset_pointer(const char* s);
@@ -66,7 +66,7 @@ private:
     char32_t        m_next;
     const char*     m_chr_ptr;
     const char*     m_chr_end;
-    int32_t         m_chr_wcwidth = 0;
+    int16_t         m_chr_wcwidth = 0;
     bool            m_emoji = false;
 };
